#9 Create pages using routes and refactor the code.
