+++ package.json
@@ -14,6 +14,7 @@
     "react-dom": "^17.0.2",
     "react-icons": "^4.3.1",
     "react-masonry-css": "^1.0.16",
+    "react-router-dom": "^6.2.1",
     "react-scripts": "5.0.0",
     "web-vitals": "^2.1.4"
   },
