+++ src/Components/About.js
@@ -1,0 +1,11 @@
+import React from "react";
+
+const About = () => {
+	return (
+		<div>
+			<p>About us</p>
+		</div>
+	);
+};
+
+export default About;
