+++ src/Components/Dashboard.js
@@ -1,0 +1,11 @@
+import React from "react";
+
+const Dashboard = () => {
+	return (
+		<div>
+			<p>Dashboard</p>
+		</div>
+	);
+};
+
+export default Dashboard;
